Refactor logic for booting a file into it's own module
